move insertCitation fn
